# Conflicts:
#	carina-aws-s3/src/main/java/com/qaprosoft/amazon/AmazonS3Manager.java
#	carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/core/factory/DriverFactory.java
#	carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/listener/DriverListener.java
#	carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/listener/IDriverCommandListener.java
#	carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/listener/ZebrunnerArtifactListener.java

+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/utils/android/IAndroidUtils.java
@@ -130,8 +130,7 @@ public interface IAndroidUtils extends IMobileUtils {
         int height = size.getHeight();
         int width = size.getWidth();
 
-        PointOption<?> option = PointOption.point(Double.valueOf(width * 0.915).intValue(),
-                Double.valueOf(height * 0.945).intValue());
+        PointOption<?> option = PointOption.point((int) (width * 0.915), (int) (height * 0.945));
         new TouchAction((AndroidDriver<?>) castDriver()).tap(option).perform();
     }
